Log warning if tempdir S3 bucket lacks an object lifecycle configuration
This patch adds automatic checks which warn users if `tempdir` points to a S3 bucket that lacks an object lifecycle configuration. The goal of this check is to make sure that users realize that these files will need to be manually cleaned up.

I also performed a bit of README cleanup and updated the build file to mark the AWS SDK as a required dependency.

Author: Josh Rosen <joshrosen@databricks.com>
Author: Josh Rosen <rosenville@gmail.com>

Closes #64 from JoshRosen/s3-lifecycle.