Refactor of AWS credentials handling
This patch makes several changes to how `spark-redshift` handles S3 credentials.

The motivation for this patch is a bug where we discovered that credentials were not being set properly for writes. This bug did not occur in our integration tests due to HDFS FileSystem caching behavior: an earlier test ended up creating a FileSystem with valid credentials and that cahced FileSystem was re-used when creating FileSystems in subsequent tests.

Unfortunately, `spark-avro` does not expose a mechanism for passing S3 credentials when writing files, so our only options for passing credentials for writes are to either encode them into the `s3n://` URI or to mutate `SparkContext`'s default `hadoopConfiguration`. Neither of these options are great:

- Due to Hadoop bugs, the URI-based approach will not work for AWS secret keys that contain backslashes.
- Mutating `hadoopConfiguration` is a bad idea because it's a confusing modification of global shared state. If the end user wants to modify this configuration, that's fine, but I feel strongly that libraries should not implicitly modify this configuration. Mutating `hadoopConfiguration` will also prevent users from using multiple sets of AWS keys in the same job, since different modifciations will overwrite each other.

It's non-trivial to extend `spark-avro` with a Hadoop Configuration-passing mechanism because `spark-avro` uses Spark's `HadoopFsRelation`, which hardcodes a dependency on `SparkContext.hadoopConf`.

Given these limitations, this patch modifies `spark-redshift` to use the following approach to handling S3 configurations:

- Remove the configuration parameters for setting AWS access keys, secret keys, and security tokens.
- Rely on the configurations to be provided via sources that are external to this library and its configuration. The Hadoop S3 FileSystems will receive their credentials via system properties, environment variables, URIs, or explicit configuration. If credentials are encoded into a S3 URI then those credentials take precedence over credentials read others mechanism.
- In order to determine which credentials to pass to Redshift, `spark-redshift` will load the credentials from the URI, Hadoop Configuration, or EC2 instance role according to the same rules / precedence used by Hadoop.

There are two corner-cases that are worth mentioning:

- If EC2 IAM authentication is being used, then `spark-redshift` still requires a set of temporary keys and a session token to pass to Redshift. In this case, we still require the user to pass these credentials to us via a set of `sts_` configs ("sts" stands for "security token service", which is an AWS feature).
- If someone is exclusively using the library through the SQL interface (such as over a JDBC / Thriftserver connection to Spark), then the only way to specify S3 credentials is by encoding them in the URI. This precludes the use of secret keys that contain backslashes. Note that the old approach of having configuration parameters for these keys did not solve this problem because we still ended up having to pass the user-provided keys into Hadoop and the only way to do so without modifying the global `hadoopConf` was to re-encode them into the URI (see discussion at #60 for more details).

Finally, there's one key point that I'd like to emphasize: if future versions of Spark and `spark-avro` have a more flexible strategy for handling Hadoop configurations, then it will always be possible to re-introduce _new_ configuration options for setting AWS keys. However, if we were to ship those options today, while we lack the support to implement consistent semantics for them, then we might be painting ourselves into an API compatibility corner.

**TL;DR**: we should be conservative in what APIs we commit to; let's start by only supporting AWS credential configuration mechanisms that have well-understood semantics, since we can always extend things in an additive manner in future releases.

Author: Josh Rosen <joshrosen@databricks.com>
Author: Michael Armbrust <michael@databricks.com>

Closes #67 from JoshRosen/s3-credentials-fixes.