Quote column names in JDBC schemaString
This patch modifies `JDBCWrapper.schemaString` to wrap column names in quotes, which is necessary in order to allow us to create tables with columns whose names are reserved words or which contain spaces. This fixes #80.

Note that, by itself, this patch does not enable full support for creating Redshift tables with column names that contain spaces; we are currently constrained by Avro's schema validation rules (see #84).

Author: Josh Rosen <joshrosen@databricks.com>

Closes #85 from JoshRosen/column-name-escaping.