Automatically use Redshift 4.0 or 4.1 JDBC driver, depending on which is installed
This patch refactors the driver-loading code so that it automatically uses either the Redshift JDBC 4.0 or 4.1 drivers, depending on which is installed. I also added support for automatically supplying an appropriate default JDBC driver class name when the `postgres://` subprotocol is used.

I tested all of the error-handling corner-cases manually.

Fixes #83.

Author: Josh Rosen <joshrosen@databricks.com>

Closes #90 from JoshRosen/auto-configure-jdbc-driver-class.