Fix bug in overwrite when using schema-qualified table names
This patch fixes #97, an issue where using `SaveMode.Overwrite` with schema-qualified table names (such as `my_schema.my_table`) would lead to "Invalid operation: syntax error at or near ".";" messages from Redshift.

The problem is that the `ALTER TABLE ... RENAME` command only accepts unqualified table names as the new names, so we need to strip off the schema prior to performing the renaming.

In order to keep this code clean, this patch introduces a new `TableName` case class which manages all of the escaping and parsing logic for table names, then uses that class to fix the Overwrite issue.

Author: Josh Rosen <joshrosen@databricks.com>

Closes #102 from JoshRosen/fix-overwrite-with-schema-qualified-table-names.