Fix thread-safety issues related to SimpleDateFormat and DecimalFormat
Java's SimpleDateFormat and DecimalFormat classes are not thread-safe, which caused problems because `spark-redshift` was using global instances of these classes.

There are multiple possible strategies for fixing this. To work around the SimpleDateFormat issue, we could have switched to using Joda Time or `commons-lang`, both of which have thread-safe date formatters. However, I'm a bit wary of introducing another dependency, particularly one which might conflict with user libraries. In addition, switching the date/time library still won't solve the decimal format problem.

The approach implemented in this patch is to create new `*Format` instances in every task as part of creating converter functions. While working on this, I took the opportunity to perform a performance-related refactoring where the read-path conversion operates by looping over an array of functions rather than using a match statement inside of the loop. This idiom is used throughout Spark SQL's own internals (see CatalystTypeConverters).

Fixes #107.

Author: Josh Rosen <joshrosen@databricks.com>

Closes #108 from JoshRosen/fix-formatter-thread-safety-problems.