Restore compatibility with spark-avro 1.0.0
This patch addresses an issue where `spark-redshift` would run into "Mandatory url is not present in manifest file" errors when saving data back to Redshift if a pre-2.0.0 version of `spark-avro` was used at runtime. This problem should not arise for most users of `spark-redshift`, since the proper version of `spark-avro` should automatically be pulled in via Maven or Ivy; the goal of this patch is to provide compatibility for users who cannot upgrade `spark-avro` for other compatibility-related reasons.

**Cause of the write path bug**: In order to fix a crash that could occur when writing tables containing empty partitions, #99 modified the write path to use manifest files that instruct Redshift to load only the non-empty partitions' Avro files. The code that handled processing of filenames when generating the manifest made assumptions about filenames that hold for spark-avro 2.0.0+ but not for spark-avro 1.0.0.

**Solution**: use a different method to build the correct list of part file names when constructing the manifest.

Fixes #111.

Author: Josh Rosen <joshrosen@databricks.com>

Closes #114 from JoshRosen/compatibility-with-spark-avro-1.0.