Have AWSCredentialsUtils.load() perform the check for STS credentials
This patch moves the check for STS credentials into the `AWSCredentials.load()` call itself. The motivation for this is a single call-site where we forgot to do a `.getOrElse()`, which caused the STS credentials to be ignored. This broke `spark-redshift` for users who authenticated to AWS only via IAM instance profiles.

Author: Josh Rosen <joshrosen@databricks.com>

Closes #159 from JoshRosen/uniform-handling-of-sts-credentials.