Quote dates and timestamps constants when pushing down filters to Redshift
This patch fixes an issue which could lead to incorrect results for queries which contained filters with date or timestamp constants. We need to quote the date and timestamp literals in order to prevent them from being interpreted as arithmetic expressions or producing syntax errors. For more details, see #152.

Fixes #152.

/cc gregrahn, who reported this issue, and marmbrus for review.

Author: Josh Rosen <joshrosen@databricks.com>

Closes #156 from JoshRosen/quote-dates-and-timestamps-in-filter-pushdown.