Allow Redshift user and password to be passed as options
Per https://github.com/databricks/spark-redshift/issues/132, this PR adds user and password options.  If these options are present and the user/password are still in the URL, an exception is thrown with a clear error message.

Unit tests are still pending - maybe it's better to check for ambiguous parameters in the Parameter class instead, to make it easier to test?  Glad for any other feedback as well!

Author: Tristan Reid <tristan@datascience.com>

Closes #162 from tristanreid/master.