Fix flakiness in "save with all empty partitions" test
The "save with all empty partitions" test is sometimes flaky in Travis. It seems that issuing a `CREATE TABLE` statement from one connection and then immediately checking for existence from a different connection can occasionally fail; to work around this, I added a sleep and re-try in the test assertions.

I also fixed an unrelated test cleanup issue in one of the IAM role tests.

Fixes #146.

Author: Josh Rosen <joshrosen@databricks.com>

Closes #229 from JoshRosen/fix-flaky-test.