Simplify use of transactions when performing overwrites and when creating new tables
This patch aims to significantly simplify our use of transactions when performing overwrites and when creating new tables; it addresses a few shortcomings in the existing code: previously, we didn't issue the `CREATE TABLE` and `COPY` commands as part of the same transaction when creating new tables and when performing overwrites we created a temporary table outside of a transaction, opening the possibility for that table to be leaked in rare corner-cases.

To address this, this patch refactors the main write logic so that it uses a JDBC connection with auto-commit disabled. Explicitly issuing `commit()` lets us significantly simplify the code and reduces duplication between the branches for handling different SaveModes.

One important change: previously, `usestagingtable` would literally create a separate staging table, but now we will simply delete the existing table and re-create it in a transaction; this frees us of the burden of having to find a unique name for the staging table. This should be safe due to Redshift's serializable snapshot isolation semantics. When `usestagingtable` is disabled, this now means that we simply issue an extra `commit()` after the original table's deletion so that the original table's resources are freed (which lets us write the new rows without having to maintain both the old and new tables at the same time).

/cc jaley, who wrote the original staging-table-handling code; it would be good if you could confirm whether the changes here will undermine the performance benefits / semantics of the old `usestagingtable=false` path.

Fixes #153 and #154.

Author: Josh Rosen <joshrosen@databricks.com>

Closes #157 from JoshRosen/use-transactions-in-more-places.