Update AWSCredentialUtils to be uniform between URI schemes.
Noticed that the way credentials were being obtained for `s3`/`s3n` were different than `s3a`. There shouldn't be different methods of getting credentials for the URI schemes as they should still be able to authenticate through the same means.

This was causing me headaches as I tried to configure a spark job on EMR, but kept getting credential errors. After digging in the code and seeing the difference in how credentials are obtained, I tried using the `s3a` URI scheme for the temp path, which finally worked. Additionally the fallback to using `InstanceProfileCredentialsProvider` in the `s3a` method allowed me to not explicitly mess with credentials and just have the library "just work".

Author: a-morales <antonio.morales@originate.com>

Closes #277 from a-morales/update-aws-credentials-utils.