Use PreparedStatement.getMetaData() to retrieve Redshift query schemas
Previously, this library issued a query against Redshift and inspected the ResultSet's metadata to determine the column names and types. Unfortunately, this approach does not work in a few corner-cases, including retrieving the schema of queries which use the `approximate count(distinct ...)` construct.

This patch fixes this issue by modifying our code to call `getMetaData()` on the prepared statement rather than running the query.

Fixes #300.

Author: Josh Rosen <joshrosen@databricks.com>

Closes #301 from JoshRosen/approx-count-distinct-read.