Fix decimal precision loss when reading the results of a Redshift query
This patch addresses a somewhat obscure corner-case bug where we would lose decimal precision when using `query` to load the results of a Redshift query. For an example of this, see #203.

The root problem is that the `SELECT * FROM $table WHERE 1=0` query that we were using to query Redshift schemas seems to affect the precision of decimal columns. If I replace the `WHERE 1=0` with a `LIMIT 1` (since `LIMIT 0` has the same problem), then Redshift reports the correct precision.

Fixes #203.

Author: Josh Rosen <joshrosen@databricks.com>

Closes #204 from JoshRosen/issue203.