Disabling EMRFS CSE during Manifest Generation
When data is encrypted in S3 and a COPY command is invoked, it's expected
that the manifest is not encrypted and is in plain-text form.

Encryption on EMR through EMRFS is controlled by a Hadoop option (fs.enable.cse).
Once set, all data that goes through the file system will be encrypted.

This commit adds an exception around generation of manifest files so that even
if the encryption option is set, the manifest file created on S3 is not encrypted.

This enables Redshift to read the manifest and ingest the data even for cases where
data is encrypted on the client side with a symmetric encryption key.