Switching UNLOAD to New Authorization
Switching the UNLOAD statement to no longer use the WITH CREDENTIALS
method and instead rely on explicitly passing the role, access key,
secret key, session token, etc.

Generally speaking this is a more flexible way of passing credentials,
though for UNLOAD it doesn't make much difference. This change is
pursued to achieve consistency with the COPY command. In COPY command,
this change is necessary to enable copy of client-side encrypted
data with Redshift.

http://docs.aws.amazon.com/redshift/latest/dg/copy-parameters-authorization.html